# ⚡️ Implement non-blocking screenshot uploads